+++ xbmc/guilib/GUIMessage.cpp
@@ -12,7 +12,7 @@
 
 std::string CGUIMessage::empty_string;
 
-CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int param1, int param2)
+CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int64_t param1, int64_t param2)
 {
   m_message = msg;
   m_senderID = senderID;
@@ -22,7 +22,7 @@ CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int param1, int p
   m_pointer = NULL;
 }
 
-CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int param1, int param2, CFileItemList *item)
+CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int64_t param1, int64_t param2, CFileItemList *item)
 {
   m_message = msg;
   m_senderID = senderID;
@@ -33,7 +33,7 @@ CGUIMessage::CGUIMessage(int msg, int senderID, int controlID, int param1, int p
 }
 
 CGUIMessage::CGUIMessage(
-    int msg, int senderID, int controlID, int param1, int param2, const CGUIListItemPtr& item)
+    int msg, int senderID, int controlID, int64_t param1, int64_t param2, const CGUIListItemPtr& item)
   : m_item(item)
 {
   m_message = msg;
@@ -69,12 +69,12 @@ CGUIListItemPtr CGUIMessage::GetItem() const
   return m_item;
 }
 
-int CGUIMessage::GetParam1() const
+int64_t CGUIMessage::GetParam1() const
 {
   return m_param1;
 }
 
-int CGUIMessage::GetParam2() const
+int64_t CGUIMessage::GetParam2() const
 {
   return m_param2;
 }
@@ -86,12 +86,12 @@ int CGUIMessage::GetSenderId() const
 
 CGUIMessage& CGUIMessage::operator = (const CGUIMessage& msg) = default;
 
-void CGUIMessage::SetParam1(int param1)
+void CGUIMessage::SetParam1(int64_t param1)
 {
   m_param1 = param1;
 }
 
-void CGUIMessage::SetParam2(int param2)
+void CGUIMessage::SetParam2(int64_t param2)
 {
   m_param2 = param2;
 }
